+++ OzzTradeDiary.NET/OzzTradeDiary.AppInfra/DesignTime/TradeMockRepository.cs
@@ -166,5 +166,10 @@ namespace TD.AppInfra.DesignTime
         {
             throw new NotImplementedException();
         }
+
+        public Task SaveTradeImagesAsync(Trade trade, bool updateTrade = true)
+        {
+            throw new NotImplementedException();
+        }
     }
 }