+++ OzzTradeDiary.NET/TD.RepositoryContracts/ITradeRepository.part.cs
@@ -9,5 +9,6 @@ namespace TD.RepositoryContracts
         Task SaveEntryOrdersAsync(Trade trade, bool updateTrade = true);
         Task SaveStopLossOrdersAsync(Trade trade, bool updateTrade = true);
         Task SaveTakeProfitOrdersAsync(Trade trade, bool updateTrade = true);
+        Task SaveTradeImagesAsync(Trade trade, bool updateTrade = true);
     }
 }