+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [0.2.2] - 2026-06-05
+
+### Added
+- Added `TradeImageCreateView`/`EditView` and corresponding ViewModels
+
+### Changed
+- Implemented `TradeImage` create/edit/delete commands and dialogs in WPF
+- Extended `TradeHistoryVM` for `TradeImage` management and selection
+- Added `SaveTradeImagesAsync` to `ITradeRepository` and implementations
+- Updated `IWindowDialogService` for `TradeImage` dialogs
+- Bumped version to 0.2.2 and updated changelog
+
 ## [0.2.1] - 2026-06-04
 
 ### Added